Contractors Added, datagrid created, database included in ContractorsServiceDataGrid

+++ Buyer.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -8,7 +9,7 @@ namespace dksApp
 {
     public class Buyer
     {
-        public uint IdBuyer { get; set; }
+        public long IdBuyer { get; set; }
         public string? BuyerName { get; set; }
         public string? BuyerStreet { get; set; }
         public string? BuyerCity { get; set; }
